refactor: Fix logging of parameters by LLM
